refactor: split test_localize.py into three focused files
