+++ tests/test_localize.py
@@ -1,9 +1,7 @@
-"""Tests for the localization tool (localize_compose, _download_https)."""
+"""Tests for the core localization tool (download, skip, compose, dedup)."""
 
-import http.client
 import io
 import os
-import urllib.request
 from unittest.mock import MagicMock, patch
 
 import pytest
@@ -12,16 +10,9 @@ from productmd.images import Image, Images
 from productmd.localize import (
     HttpTask,
     LocalizeResult,
-    OciTask,
-    _SafeRedirectHandler,
-    _build_auth_header,
     _deduplicate_http_tasks,
     _download_https,
-    _download_single_oci,
-    _get_netrc_auth_header,
     _should_skip,
-    _should_skip_oci,
-    _validate_credential,
     localize_compose,
 )
 from productmd.location import FileEntry, Location
@@ -615,462 +606,6 @@ class TestLocalizeCompose:
 
 
 # ---------------------------------------------------------------------------
-# Tests: OCI-specific skip and download logic
-# ---------------------------------------------------------------------------
-
-
-class TestShouldSkipOci:
-    """Tests for the _should_skip_oci function."""
-
-    def test_skip_oci_contents_all_exist(self, tmp_path):
-        """Test skip when all content files exist with valid checksums."""
-        from productmd.location import compute_checksum as cc
-
-        dest_dir = str(tmp_path / "images")
-        os.makedirs(os.path.join(dest_dir, "pxeboot"))
-
-        # Create files
-        vmlinuz = os.path.join(dest_dir, "pxeboot", "vmlinuz")
-        with open(vmlinuz, "wb") as f:
-            f.write(b"kernel content")
-        checksum = cc(vmlinuz, "sha256")
-
-        contents = [
-            FileEntry(file="pxeboot/vmlinuz", size=14, checksum=checksum, layer_digest="sha256:" + "a" * 64),
-        ]
-        task = OciTask(
-            oci_url="oci://registry/repo@sha256:" + "b" * 64,
-            dest_dir=dest_dir,
-            contents=contents,
-            location=Location(url="oci://registry/repo@sha256:" + "b" * 64, local_path="images"),
-            rel_path="images/pxeboot/vmlinuz",
-        )
-
-        assert _should_skip_oci(task, verify_checksums=True)
-
-    def test_no_skip_oci_contents_missing_file(self, tmp_path):
-        """Test no skip when a content file is missing."""
-        dest_dir = str(tmp_path / "images")
-        os.makedirs(dest_dir)
-
-        contents = [
-            FileEntry(file="pxeboot/vmlinuz", size=14, checksum=None, layer_digest="sha256:" + "a" * 64),
-        ]
-        task = OciTask(
-            oci_url="oci://registry/repo@sha256:" + "b" * 64,
-            dest_dir=dest_dir,
-            contents=contents,
-            location=Location(url="oci://registry/repo@sha256:" + "b" * 64, local_path="images"),
-            rel_path="images/pxeboot/vmlinuz",
-        )
-
-        assert not _should_skip_oci(task, verify_checksums=True)
-
-    def test_no_skip_oci_contents_bad_checksum(self, tmp_path):
-        """Test no skip when a content file has wrong checksum."""
-        dest_dir = str(tmp_path / "images")
-        os.makedirs(os.path.join(dest_dir, "pxeboot"))
-
-        vmlinuz = os.path.join(dest_dir, "pxeboot", "vmlinuz")
-        with open(vmlinuz, "wb") as f:
-            f.write(b"kernel content")
-
-        contents = [
-            FileEntry(
-                file="pxeboot/vmlinuz",
-                size=14,
-                checksum="sha256:" + "f" * 64,  # wrong
-                layer_digest="sha256:" + "a" * 64,
-            ),
-        ]
-        task = OciTask(
-            oci_url="oci://registry/repo@sha256:" + "b" * 64,
-            dest_dir=dest_dir,
-            contents=contents,
-            location=Location(url="oci://registry/repo@sha256:" + "b" * 64, local_path="images"),
-            rel_path="images/pxeboot/vmlinuz",
-        )
-
-        assert not _should_skip_oci(task, verify_checksums=True)
-
-    def test_skip_oci_contents_no_verify(self, tmp_path):
-        """Test skip when all content files exist and verify_checksums=False."""
-        dest_dir = str(tmp_path / "images")
-        os.makedirs(os.path.join(dest_dir, "pxeboot"))
-
-        vmlinuz = os.path.join(dest_dir, "pxeboot", "vmlinuz")
-        with open(vmlinuz, "wb") as f:
-            f.write(b"kernel content")
-
-        contents = [
-            FileEntry(
-                file="pxeboot/vmlinuz",
-                size=14,
-                checksum="sha256:" + "f" * 64,  # wrong but not checked
-                layer_digest="sha256:" + "a" * 64,
-            ),
-        ]
-        task = OciTask(
-            oci_url="oci://registry/repo@sha256:" + "b" * 64,
-            dest_dir=dest_dir,
-            contents=contents,
-            location=Location(url="oci://registry/repo@sha256:" + "b" * 64, local_path="images"),
-            rel_path="images/pxeboot/vmlinuz",
-        )
-
-        # Not verified, so existence alone is enough
-        assert _should_skip_oci(task, verify_checksums=False)
-
-    def test_skip_oci_simple_existing_file(self, tmp_path):
-        """Test skip for simple OCI (no contents) when file exists."""
-        # For simple OCI, dest_dir is the compose root and the file
-        # is at dest_dir/local_path
-        dest_dir = str(tmp_path / "compose")
-        os.makedirs(dest_dir)
-        file_path = os.path.join(dest_dir, "output.iso")
-        with open(file_path, "wb") as f:
-            f.write(b"iso content")
-
-        task = OciTask(
-            oci_url="oci://registry/repo@sha256:" + "b" * 64,
-            dest_dir=dest_dir,
-            contents=[],
-            location=Location(url="oci://registry/repo@sha256:" + "b" * 64, local_path="output.iso"),
-            rel_path="output.iso",
-        )
-
-        assert _should_skip_oci(task, verify_checksums=False)
-
-
-class TestOciLocalizeIntegration:
-    """Integration-style tests for OCI localization within localize_compose."""
-
-    @patch("productmd.oci.get_downloader")
-    @patch("productmd.oci.HAS_ORAS", True)
-    def test_oci_download_error_fail_fast(self, mock_get_downloader, tmp_path):
-        """Test that OCI download errors respect fail_fast."""
-        mock_downloader = MagicMock()
-        mock_downloader.download_and_extract.side_effect = RuntimeError("registry unreachable")
-        mock_get_downloader.return_value = mock_downloader
-
-        im = Images()
-        im.header.version = "2.0"
-        im.compose.id = "Test-1.0-20260204.0"
-        im.compose.type = "production"
-        im.compose.date = "20260204"
-        im.compose.respin = 0
-        im.output_version = VERSION_2_0
-
-        img = _make_image(im, "Server/x86_64/iso/boot.iso", 512, "a" * 64)
-        img.location = Location(
-            url="oci://quay.io/fedora/server:41-x86_64@sha256:" + "a" * 64,
-            size=512,
-            checksum="sha256:" + "a" * 64,
-            local_path="Server/x86_64/iso/boot.iso",
-        )
-        im.add("Server", "x86_64", img)
-
-        result = localize_compose(
-            output_dir=str(tmp_path / "output"),
-            images=im,
-            fail_fast=True,
-            verify_checksums=False,
-        )
-
-        assert result.failed == 1
-        assert result.downloaded == 0
-        assert len(result.errors) == 1
-        assert "registry unreachable" in str(result.errors[0][1])
-
-    @patch("productmd.oci.get_downloader")
-    @patch("productmd.oci.HAS_ORAS", True)
-    def test_oci_progress_events(self, mock_get_downloader, tmp_path):
-        """Test that OCI downloads emit progress events."""
-        mock_downloader = MagicMock()
-        mock_get_downloader.return_value = mock_downloader
-
-        im = Images()
-        im.header.version = "2.0"
-        im.compose.id = "Test-1.0-20260204.0"
-        im.compose.type = "production"
-        im.compose.date = "20260204"
-        im.compose.respin = 0
-        im.output_version = VERSION_2_0
-
-        img = _make_image(im, "Server/x86_64/iso/boot.iso", 512, "a" * 64)
-        img.location = Location(
-            url="oci://quay.io/fedora/server:41-x86_64@sha256:" + "a" * 64,
-            size=512,
-            checksum="sha256:" + "a" * 64,
-            local_path="Server/x86_64/iso/boot.iso",
-        )
-        im.add("Server", "x86_64", img)
-
-        events = []
-        localize_compose(
-            output_dir=str(tmp_path / "output"),
-            images=im,
-            verify_checksums=False,
-            progress_callback=events.append,
-        )
-
-        event_types = [e.event_type for e in events]
-        assert "start" in event_types
-        assert "complete" in event_types
-
-    @patch("productmd.oci.get_downloader")
-    @patch("productmd.oci.HAS_ORAS", True)
-    @patch("productmd.localize.urllib.request.urlopen")
-    def test_mixed_http_and_oci_downloads(self, mock_urlopen, mock_get_downloader, tmp_path):
-        """Test that a compose with both HTTP and OCI URLs works."""
-        mock_urlopen.return_value = _mock_urlopen(b"http content")
-        mock_downloader = MagicMock()
-        mock_get_downloader.return_value = mock_downloader
-
-        im = Images()
-        im.header.version = "2.0"
-        im.compose.id = "Test-1.0-20260204.0"
-        im.compose.type = "production"
-        im.compose.date = "20260204"
-        im.compose.respin = 0
-        im.output_version = VERSION_2_0
-
-        # HTTP image
-        img_http = _make_image(im, "Server/x86_64/iso/boot.iso", 512, "a" * 64)
-        img_http.location = Location(
-            url="https://cdn.example.com/Server/x86_64/iso/boot.iso",
-            size=512,
-            checksum="sha256:" + "a" * 64,
-            local_path="Server/x86_64/iso/boot.iso",
-        )
-        im.add("Server", "x86_64", img_http)
-
-        # OCI image
-        img_oci = _make_image(im, "Server/aarch64/iso/boot.iso", 1024, "b" * 64)
-        img_oci.arch = "aarch64"
-        img_oci.location = Location(
-            url="oci://quay.io/fedora/server:41-aarch64@sha256:" + "b" * 64,
-            size=1024,
-            checksum="sha256:" + "b" * 64,
-            local_path="Server/aarch64/iso/boot.iso",
-        )
-        im.add("Server", "aarch64", img_oci)
-
-        result = localize_compose(
-            output_dir=str(tmp_path / "output"),
-            images=im,
-            parallel_downloads=1,
-            verify_checksums=False,
-            retries=0,
-        )
-
-        # 1 HTTP + 1 OCI
-        assert result.downloaded == 2
-        assert result.failed == 0
-        mock_urlopen.assert_called_once()
-        mock_downloader.download_and_extract.assert_called_once()
-
-
-# ---------------------------------------------------------------------------
-# Tests: Parallel OCI downloads
-# ---------------------------------------------------------------------------
-
-
-class TestOciParallelDownloads:
-    """Tests for parallel OCI download execution."""
-
-    @patch("productmd.oci.get_downloader")
-    @patch("productmd.oci.HAS_ORAS", True)
-    def test_parallel_multiple_oci_tasks(self, mock_get_downloader, tmp_path):
-        """Test that multiple OCI tasks are all downloaded in parallel mode."""
-        mock_downloader = MagicMock()
-        mock_get_downloader.return_value = mock_downloader
-
-        im = Images()
-        im.header.version = "2.0"
-        im.compose.id = "Test-1.0-20260204.0"
-        im.compose.type = "production"
-        im.compose.date = "20260204"
-        im.compose.respin = 0
-        im.output_version = VERSION_2_0
-
-        # Create 3 OCI images on different arches
-        for arch, hex_char in [("x86_64", "a"), ("aarch64", "b"), ("s390x", "c")]:
-            img = _make_image(im, f"Server/{arch}/iso/boot.iso", 512, hex_char * 64)
-            img.arch = arch
-            img.location = Location(
-                url=f"oci://quay.io/fedora/server:41-{arch}@sha256:" + hex_char * 64,
-                size=512,
-                checksum="sha256:" + hex_char * 64,
-                local_path=f"Server/{arch}/iso/boot.iso",
-            )
-            im.add("Server", arch, img)
-
-        result = localize_compose(
-            output_dir=str(tmp_path / "output"),
-            images=im,
-            parallel_downloads=3,
-            verify_checksums=False,
-        )
-
-        assert result.downloaded == 3
-        assert result.failed == 0
-        # Each task creates its own downloader (thread-safety)
-        assert mock_get_downloader.call_count == 3
-        assert mock_downloader.download_and_extract.call_count == 3
-
-    @patch("productmd.oci.get_downloader")
-    @patch("productmd.oci.HAS_ORAS", True)
-    def test_parallel_oci_fail_fast_stops(self, mock_get_downloader, tmp_path):
-        """Test that fail_fast cancels remaining parallel OCI tasks on error."""
-        mock_downloader = MagicMock()
-        mock_downloader.download_and_extract.side_effect = RuntimeError("pull failed")
-        mock_get_downloader.return_value = mock_downloader
-
-        im = Images()
-        im.header.version = "2.0"
-        im.compose.id = "Test-1.0-20260204.0"
-        im.compose.type = "production"
-        im.compose.date = "20260204"
-        im.compose.respin = 0
-        im.output_version = VERSION_2_0
-
-        for arch, hex_char in [("x86_64", "a"), ("aarch64", "b"), ("s390x", "c")]:
-            img = _make_image(im, f"Server/{arch}/iso/boot.iso", 512, hex_char * 64)
-            img.arch = arch
-            img.location = Location(
-                url=f"oci://quay.io/fedora/server:41-{arch}@sha256:" + hex_char * 64,
-                size=512,
-                checksum="sha256:" + hex_char * 64,
-                local_path=f"Server/{arch}/iso/boot.iso",
-            )
-            im.add("Server", arch, img)
-
-        result = localize_compose(
-            output_dir=str(tmp_path / "output"),
-            images=im,
-            parallel_downloads=2,
-            fail_fast=True,
-            verify_checksums=False,
-        )
-
-        assert result.downloaded == 0
-        assert result.failed >= 1
-        assert len(result.errors) >= 1
-
-    @patch("productmd.oci.get_downloader")
-    @patch("productmd.oci.HAS_ORAS", True)
-    def test_parallel_oci_no_fail_fast_continues(self, mock_get_downloader, tmp_path):
-        """Test that fail_fast=False continues after OCI errors in parallel mode."""
-        call_count = 0
-
-        def side_effect(**kwargs):
-            nonlocal call_count
-            call_count += 1
-            if call_count == 1:
-                raise RuntimeError("first pull failed")
-            # Subsequent calls succeed
-
-        mock_downloader = MagicMock()
-        mock_downloader.download_and_extract.side_effect = side_effect
-        mock_get_downloader.return_value = mock_downloader
-
-        im = Images()
-        im.header.version = "2.0"
-        im.compose.id = "Test-1.0-20260204.0"
-        im.compose.type = "production"
-        im.compose.date = "20260204"
-        im.compose.respin = 0
-        im.output_version = VERSION_2_0
-
-        for arch, hex_char in [("x86_64", "a"), ("aarch64", "b")]:
-            img = _make_image(im, f"Server/{arch}/iso/boot.iso", 512, hex_char * 64)
-            img.arch = arch
-            img.location = Location(
-                url=f"oci://quay.io/fedora/server:41-{arch}@sha256:" + hex_char * 64,
-                size=512,
-                checksum="sha256:" + hex_char * 64,
-                local_path=f"Server/{arch}/iso/boot.iso",
-            )
-            im.add("Server", arch, img)
-
-        result = localize_compose(
-            output_dir=str(tmp_path / "output"),
-            images=im,
-            parallel_downloads=2,
-            fail_fast=False,
-            verify_checksums=False,
-        )
-
-        assert result.downloaded == 1
-        assert result.failed == 1
-        assert len(result.errors) == 1
-
-    @patch("productmd.oci.get_downloader")
-    @patch("productmd.oci.HAS_ORAS", True)
-    def test_sequential_oci_with_parallel_1(self, mock_get_downloader, tmp_path):
-        """Test that parallel_downloads=1 uses sequential OCI path."""
-        mock_downloader = MagicMock()
-        mock_get_downloader.return_value = mock_downloader
-
-        im = Images()
-        im.header.version = "2.0"
-        im.compose.id = "Test-1.0-20260204.0"
-        im.compose.type = "production"
-        im.compose.date = "20260204"
-        im.compose.respin = 0
-        im.output_version = VERSION_2_0
-
-        for arch, hex_char in [("x86_64", "a"), ("aarch64", "b")]:
-            img = _make_image(im, f"Server/{arch}/iso/boot.iso", 512, hex_char * 64)
-            img.arch = arch
-            img.location = Location(
-                url=f"oci://quay.io/fedora/server:41-{arch}@sha256:" + hex_char * 64,
-                size=512,
-                checksum="sha256:" + hex_char * 64,
-                local_path=f"Server/{arch}/iso/boot.iso",
-            )
-            im.add("Server", arch, img)
-
-        result = localize_compose(
-            output_dir=str(tmp_path / "output"),
-            images=im,
-            parallel_downloads=1,
-            verify_checksums=False,
-        )
-
-        assert result.downloaded == 2
-        assert result.failed == 0
-        # Each task still creates its own downloader
-        assert mock_get_downloader.call_count == 2
-
-    @patch("productmd.oci.get_downloader")
-    @patch("productmd.oci.HAS_ORAS", True)
-    def test_download_single_oci_creates_own_downloader(self, mock_get_downloader, tmp_path):
-        """Test that _download_single_oci creates a fresh downloader per call."""
-        mock_downloader = MagicMock()
-        mock_get_downloader.return_value = mock_downloader
-
-        task = OciTask(
-            oci_url="oci://registry/repo@sha256:" + "a" * 64,
-            dest_dir=str(tmp_path / "out"),
-            contents=[],
-            location=Location(
-                url="oci://registry/repo@sha256:" + "a" * 64,
-                size=100,
-                local_path="out",
-            ),
-            rel_path="out/file",
-        )
-
-        _download_single_oci(task, progress_callback=None)
-        _download_single_oci(task, progress_callback=None)
-
-        # get_downloader called once per invocation (fresh downloader each time)
-        assert mock_get_downloader.call_count == 2
-
-
-# ---------------------------------------------------------------------------
 # Tests: HTTP task deduplication
 # ---------------------------------------------------------------------------
 
@@ -1123,519 +658,3 @@ class TestDeduplicateHttpTasks:
         )
         result = _deduplicate_http_tasks([task1, task2])
         assert len(result) == 2
-
-
-# ---------------------------------------------------------------------------
-# Tests: HTTP Authentication
-# ---------------------------------------------------------------------------
-
-
-class TestGetNetrcAuthHeader:
-    """Tests for _get_netrc_auth_header."""
-
-    def test_netrc_auth_found(self, tmp_path):
-        """Test that matching netrc credentials produce a Basic auth header."""
-        netrc_file = tmp_path / ".netrc"
-        netrc_file.write_text("machine pulp.example.com\nlogin admin\npassword secret123\n")
-
-        result = _get_netrc_auth_header("https://pulp.example.com/repo/file.rpm", str(netrc_file))
-
-        assert result is not None
-        assert result.startswith("Basic ")
-        import base64
-
-        decoded = base64.b64decode(result.split(" ", 1)[1]).decode()
-        assert decoded == "admin:secret123"
-
-    def test_netrc_host_not_found(self, tmp_path):
-        """Test that no match returns None."""
-        netrc_file = tmp_path / ".netrc"
-        netrc_file.write_text("machine other.example.com\nlogin user\npassword pass\n")
-
-        result = _get_netrc_auth_header("https://pulp.example.com/file.rpm", str(netrc_file))
-
-        assert result is None
-
-    def test_netrc_file_missing(self):
-        """Test that a missing netrc file returns None without error."""
-        result = _get_netrc_auth_header("https://pulp.example.com/file.rpm", "/nonexistent/.netrc")
-
-        assert result is None
-
-    def test_netrc_parse_error(self, tmp_path):
-        """Test that a malformed netrc file returns None without error."""
-        netrc_file = tmp_path / ".netrc"
-        netrc_file.write_text("this is not valid netrc content!!!\n~~~")
-
-        result = _get_netrc_auth_header("https://pulp.example.com/file.rpm", str(netrc_file))
-
-        assert result is None
-
-    def test_netrc_custom_file(self, tmp_path):
-        """Test that a custom netrc file path is used."""
-        custom_netrc = tmp_path / "custom-netrc"
-        custom_netrc.write_text("machine cdn.example.com\nlogin deploy\npassword s3cret\n")
-
-        result = _get_netrc_auth_header("https://cdn.example.com/boot.iso", str(custom_netrc))
-
-        assert result is not None
-        import base64
-
-        decoded = base64.b64decode(result.split(" ", 1)[1]).decode()
-        assert decoded == "deploy:s3cret"
-
-    def test_netrc_no_hostname_in_url(self):
-        """Test that a URL with no hostname returns None."""
-        result = _get_netrc_auth_header("file:///local/path", None)
-
-        assert result is None
-
-
-class TestBuildAuthHeader:
-    """Tests for _build_auth_header precedence logic."""
-
-    def test_explicit_basic_auth(self):
-        """Test that explicit username/password produce a Basic header."""
-        result = _build_auth_header("https://example.com/file", username="user", password="pass")
-
-        assert result is not None
-        assert result.startswith("Basic ")
-        import base64
-
-        decoded = base64.b64decode(result.split(" ", 1)[1]).decode()
-        assert decoded == "user:pass"
-
-    def test_bearer_token(self):
-        """Test that a token produces a Bearer header."""
-        result = _build_auth_header("https://example.com/file", token="my-jwt-token")
-
-        assert result == "Bearer my-jwt-token"
-
-    def test_token_overrides_basic(self):
-        """Test that token takes precedence over username/password."""
-        result = _build_auth_header(
-            "https://example.com/file",
-            username="user",
-            password="pass",
-            token="my-token",
-        )
-
-        assert result == "Bearer my-token"
-
-    def test_token_overrides_netrc(self, tmp_path):
-        """Test that token takes precedence over netrc credentials."""
-        netrc_file = tmp_path / ".netrc"
-        netrc_file.write_text("machine example.com\nlogin admin\npassword secret\n")
-
-        result = _build_auth_header(
-            "https://example.com/file",
-            token="my-token",
-            netrc_file=str(netrc_file),
-        )
-
-        assert result == "Bearer my-token"
-
-    def test_explicit_basic_overrides_netrc(self, tmp_path):
-        """Test that explicit credentials override netrc."""
-        netrc_file = tmp_path / ".netrc"
-        netrc_file.write_text("machine example.com\nlogin netrc-user\npassword netrc-pass\n")
-
-        result = _build_auth_header(
-            "https://example.com/file",
-            username="explicit-user",
-            password="explicit-pass",
-            netrc_file=str(netrc_file),
-        )
-
-        assert result is not None
-        assert result.startswith("Basic ")
-        import base64
-
-        decoded = base64.b64decode(result.split(" ", 1)[1]).decode()
-        assert decoded == "explicit-user:explicit-pass"
-
-    def test_falls_back_to_netrc(self, tmp_path):
-        """Test that netrc is used when no explicit credentials are given."""
-        netrc_file = tmp_path / ".netrc"
-        netrc_file.write_text("machine example.com\nlogin netrc-user\npassword netrc-pass\n")
-
-        result = _build_auth_header("https://example.com/file", netrc_file=str(netrc_file))
-
-        assert result is not None
-        import base64
-
-        decoded = base64.b64decode(result.split(" ", 1)[1]).decode()
-        assert decoded == "netrc-user:netrc-pass"
-
-    def test_no_credentials_returns_none(self):
-        """Test that None is returned when no auth is available."""
-        result = _build_auth_header(
-            "https://example.com/file",
-            netrc_file="/nonexistent/.netrc",
-        )
-
-        assert result is None
-
-
-class TestDownloadHttpsAuth:
-    """Tests for authentication in _download_https."""
-
-    @patch("productmd.localize.urllib.request.urlopen")
-    def test_download_sends_basic_auth_header(self, mock_urlopen, tmp_path):
-        """Test that explicit credentials add Authorization header to request."""
-        mock_urlopen.return_value = _mock_urlopen(b"content")
-        dest = str(tmp_path / "file.rpm")
-
-        _download_https(
-            "https://pulp.example.com/file.rpm",
-            dest,
-            retries=0,
-            username="admin",
-            password="secret",
-        )
-
-        req = mock_urlopen.call_args[0][0]
-        assert req.get_header("Authorization").startswith("Basic ")
-
-    @patch("productmd.localize.urllib.request.urlopen")
-    def test_download_sends_bearer_token_header(self, mock_urlopen, tmp_path):
-        """Test that a token adds a Bearer Authorization header."""
-        mock_urlopen.return_value = _mock_urlopen(b"content")
-        dest = str(tmp_path / "file.rpm")
-
-        _download_https(
-            "https://pulp.example.com/file.rpm",
-            dest,
-            retries=0,
-            token="jwt-token-here",
-        )
-
-        req = mock_urlopen.call_args[0][0]
-        assert req.get_header("Authorization") == "Bearer jwt-token-here"
-
-    @patch("productmd.localize.urllib.request.urlopen")
-    def test_download_sends_netrc_auth_header(self, mock_urlopen, tmp_path):
-        """Test that netrc credentials add Authorization header."""
-        mock_urlopen.return_value = _mock_urlopen(b"content")
-        netrc_file = tmp_path / ".netrc"
-        netrc_file.write_text("machine pulp.example.com\nlogin admin\npassword secret\n")
-        dest = str(tmp_path / "file.rpm")
-
-        _download_https(
-            "https://pulp.example.com/file.rpm",
-            dest,
-            retries=0,
-            netrc_file=str(netrc_file),
-        )
-
-        req = mock_urlopen.call_args[0][0]
-        assert req.get_header("Authorization").startswith("Basic ")
-
-    @patch("productmd.localize.urllib.request.urlopen")
-    def test_download_no_auth_by_default(self, mock_urlopen, tmp_path):
-        """Test that no Authorization header is set when no auth is configured."""
-        mock_urlopen.return_value = _mock_urlopen(b"content")
-        dest = str(tmp_path / "file.rpm")
-
-        _download_https(
-            "https://pulp.example.com/file.rpm",
-            dest,
-            retries=0,
-            netrc_file="/nonexistent/.netrc",
-        )
-
-        req = mock_urlopen.call_args[0][0]
-        assert req.get_header("Authorization") is None
-
-    @patch("productmd.localize.urllib.request.urlopen")
-    def test_localize_compose_passes_auth_credentials(self, mock_urlopen, tmp_path):
-        """Test that localize_compose passes auth params to _download_https."""
-        mock_urlopen.return_value = _mock_urlopen(b"x" * 512)
-        images = _create_images_v2()
-
-        localize_compose(
-            output_dir=str(tmp_path),
-            images=images,
-            parallel_downloads=1,
-            verify_checksums=False,
-            http_username="user",
-            http_password="pass",
-        )
-
-        req = mock_urlopen.call_args[0][0]
-        assert req.get_header("Authorization").startswith("Basic ")
-
-    @patch("productmd.localize.urllib.request.urlopen")
-    def test_localize_compose_passes_token(self, mock_urlopen, tmp_path):
-        """Test that localize_compose passes token to _download_https."""
-        mock_urlopen.return_value = _mock_urlopen(b"x" * 512)
-        images = _create_images_v2()
-
-        localize_compose(
-            output_dir=str(tmp_path),
-            images=images,
-            parallel_downloads=1,
-            verify_checksums=False,
-            http_token="my-bearer-token",
-        )
-
-        req = mock_urlopen.call_args[0][0]
-        assert req.get_header("Authorization") == "Bearer my-bearer-token"
-
-
-# ---------------------------------------------------------------------------
-# Tests: Safe Redirect Handler
-# ---------------------------------------------------------------------------
-
-
-class TestSafeRedirectHandler:
-    """Tests for _SafeRedirectHandler cross-origin auth stripping."""
-
-    def _redirect(self, handler, orig_url, new_url, auth_header=None):
-        """Helper to invoke redirect_request with minimal boilerplate."""
-        req = urllib.request.Request(orig_url)
-        if auth_header:
-            req.add_header("Authorization", auth_header)
-        headers = http.client.HTTPMessage()
-        fp = io.BytesIO(b"")
-        return handler.redirect_request(req, fp, 302, "Found", headers, new_url)
-
-    def test_strips_auth_on_cross_origin_redirect(self):
-        """Test that Authorization is removed when redirecting to a different host."""
-        handler = _SafeRedirectHandler()
-
-        new_req = self._redirect(
-            handler,
-            "https://pulp.example.com/repo/file.rpm",
-            "https://cdn.example.com/cached-file.rpm",
-            auth_header="Basic abc123",
-        )
-
-        assert new_req is not None
-        assert new_req.get_header("Authorization") is None
-
-    def test_keeps_auth_on_same_origin_redirect(self):
-        """Test that Authorization is preserved when redirecting to the same origin."""
-        handler = _SafeRedirectHandler()
-
-        new_req = self._redirect(
-            handler,
-            "https://pulp.example.com/repo/file.rpm",
-            "https://pulp.example.com/different/path.rpm",
-            auth_header="Basic abc123",
-        )
-
-        assert new_req is not None
-        assert new_req.get_header("Authorization") == "Basic abc123"
-
-    def test_strips_auth_on_scheme_downgrade(self):
-        """Test that Authorization is removed on HTTPS -> HTTP downgrade."""
-        handler = _SafeRedirectHandler()
-
-        new_req = self._redirect(
-            handler,
-            "https://pulp.example.com/file.rpm",
-            "http://pulp.example.com/file.rpm",
-            auth_header="Bearer my-token",
-        )
-
-        assert new_req is not None
-        assert new_req.get_header("Authorization") is None
-
-    def test_strips_auth_on_port_change(self):
-        """Test that Authorization is removed when the port changes."""
-        handler = _SafeRedirectHandler()
-
-        new_req = self._redirect(
-            handler,
-            "https://pulp.example.com:443/file.rpm",
-            "https://pulp.example.com:8443/file.rpm",
-            auth_header="Basic abc123",
-        )
-
-        assert new_req is not None
-        assert new_req.get_header("Authorization") is None
-
-    def test_no_auth_header_no_error(self):
-        """Test that redirects without Authorization header work fine."""
-        handler = _SafeRedirectHandler()
-
-        new_req = self._redirect(
-            handler,
-            "https://pulp.example.com/file.rpm",
-            "https://cdn.example.com/file.rpm",
-        )
-
-        assert new_req is not None
-        assert new_req.get_header("Authorization") is None
-
-    def test_keeps_auth_on_default_port_redirect(self):
-        """Test that https://host → https://host:443 is treated as same origin."""
-        handler = _SafeRedirectHandler()
-
-        new_req = self._redirect(
-            handler,
-            "https://pulp.example.com/file.rpm",
-            "https://pulp.example.com:443/file.rpm",
-            auth_header="Basic abc123",
-        )
-
-        assert new_req is not None
-        assert new_req.get_header("Authorization") == "Basic abc123"
-
-    def test_keeps_auth_on_default_http_port_redirect(self):
-        """Test that http://host → http://host:80 is treated as same origin."""
-        handler = _SafeRedirectHandler()
-
-        new_req = self._redirect(
-            handler,
-            "http://pulp.example.com/file.rpm",
-            "http://pulp.example.com:80/file.rpm",
-            auth_header="Bearer my-token",
-        )
-
-        assert new_req is not None
-        assert new_req.get_header("Authorization") == "Bearer my-token"
-
-
-# ---------------------------------------------------------------------------
-# Tests: Credential validation
-# ---------------------------------------------------------------------------
-
-
-class TestValidateCredential:
-    """Tests for _validate_credential header injection prevention."""
-
-    def test_valid_credential(self):
-        """Normal credential values pass through unchanged."""
-        assert _validate_credential("admin", "username") == "admin"
-        assert _validate_credential("secret123", "password") == "secret123"
-        assert _validate_credential("eyJhbGciOiJSUzI1NiIs", "token") == "eyJhbGciOiJSUzI1NiIs"
-
-    @pytest.mark.parametrize(
-        "value",
-        [
-            "token\nevil",
-            "token\revil",
-            "token\r\nevil-header: injected",
-        ],
-    )
-    def test_rejects_newline_characters(self, value):
-        """Credentials with CR, LF, or CRLF are rejected."""
-        with pytest.raises(ValueError, match="newline"):
-            _validate_credential(value, "token")
-
-    @pytest.mark.parametrize(
-        "kwargs",
-        [
-            {"token": "bad\ntoken"},
-            {"username": "bad\nuser", "password": "pass"},
-            {"username": "user", "password": "bad\npass"},
-        ],
-    )
-    def test_build_auth_header_rejects_bad_credentials(self, kwargs):
-        """_build_auth_header raises on any credential with newlines."""
-        with pytest.raises(ValueError, match="newline"):
-            _build_auth_header("https://example.com", **kwargs)
-
-
-# ---------------------------------------------------------------------------
-# Tests: Library-level auth validation
-# ---------------------------------------------------------------------------
-
-
-class TestLocalizeComposeAuthValidation:
-    """Tests for auth parameter validation in localize_compose."""
-
-    def test_username_without_password_raises(self):
-        """localize_compose raises ValueError when username is given without password."""
-        with pytest.raises(ValueError, match="must be provided together"):
-            localize_compose(
-                output_dir="/tmp/test",
-                http_username="admin",
-            )
-
-    def test_password_without_username_raises(self):
-        """localize_compose raises ValueError when password is given without username."""
-        with pytest.raises(ValueError, match="must be provided together"):
-            localize_compose(
-                output_dir="/tmp/test",
-                http_password="secret",
-            )
-
-    def test_token_with_username_raises(self):
-        """localize_compose raises ValueError when token and username are both given."""
-        with pytest.raises(ValueError, match="mutually exclusive"):
-            localize_compose(
-                output_dir="/tmp/test",
-                http_token="my-token",
-                http_username="admin",
-                http_password="secret",
-            )
-
-
-# ---------------------------------------------------------------------------
-# Tests: Netrc default entry
-# ---------------------------------------------------------------------------
-
-
-class TestNetrcDefaultEntry:
-    """Tests for netrc 'default' machine entry."""
-
-    def test_netrc_default_entry_matches(self, tmp_path):
-        """Test that the netrc 'default' entry matches any hostname."""
-        netrc_file = tmp_path / ".netrc"
-        netrc_file.write_text("default\nlogin fallback-user\npassword fallback-pass\n")
-
-        result = _get_netrc_auth_header("https://any-host.example.com/file.rpm", str(netrc_file))
-
-        assert result is not None
-        import base64
-
-        decoded = base64.b64decode(result.split(" ", 1)[1]).decode()
-        assert decoded == "fallback-user:fallback-pass"
-
-    def test_netrc_specific_overrides_default(self, tmp_path):
-        """Test that a specific machine entry takes precedence over default."""
-        netrc_file = tmp_path / ".netrc"
-        netrc_file.write_text(
-            "machine pulp.example.com\nlogin specific-user\npassword specific-pass\n\n"
-            "default\nlogin fallback-user\npassword fallback-pass\n"
-        )
-
-        result = _get_netrc_auth_header("https://pulp.example.com/file.rpm", str(netrc_file))
-
-        assert result is not None
-        import base64
-
-        decoded = base64.b64decode(result.split(" ", 1)[1]).decode()
-        assert decoded == "specific-user:specific-pass"
-
-
-# ---------------------------------------------------------------------------
-# Tests: Parallel downloads with auth
-# ---------------------------------------------------------------------------
-
-
-class TestParallelDownloadsAuth:
-    """Tests for auth in parallel download path."""
-
-    @patch("productmd.localize.urllib.request.urlopen")
-    def test_parallel_downloads_send_auth(self, mock_urlopen, tmp_path):
-        """Test that auth headers are sent when using parallel downloads."""
-        mock_urlopen.return_value = _mock_urlopen(b"x" * 512)
-        images = _create_images_v2()
-
-        localize_compose(
-            output_dir=str(tmp_path),
-            images=images,
-            parallel_downloads=2,
-            verify_checksums=False,
-            http_username="admin",
-            http_password="secret",
-        )
-
-        # All calls should have the auth header
-        for call in mock_urlopen.call_args_list:
-            req = call[0][0]
-            assert req.get_header("Authorization").startswith("Basic ")
